+++ wallet/src/main/java/com/bbva/wallet/services/TransactionService.java
@@ -1,4 +1,8 @@
 package com.bbva.wallet.services;
 
+import com.bbva.wallet.dtos.TransactionDto;
+import com.bbva.wallet.entities.Transactions;
+
 public interface TransactionService {
+    Transactions sendArs(TransactionDto transactionDto);
 }
